stockanalysis/requirements.txt
langchain-google-genai
google-genai
yfinance
pydantic
python-dotenv
langchain
langchain-google-genai
pydantic-ai
